third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/stroke-dasharray-expected.txt

This is a testharness.js-based test.
[FAIL] Can set 'stroke-dasharray' to a length: 0px
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'stroke-dasharray' to a length: -3.14em
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'stroke-dasharray' to a length: 3.14cm
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'stroke-dasharray' to a length: calc(0px + 0em)
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'stroke-dasharray' to a percent: 0%
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'stroke-dasharray' to a percent: -3.14%
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'stroke-dasharray' to a percent: 3.14%
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'stroke-dasharray' to a percent: calc(0% + 0%)
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'stroke-dasharray' to a number: 0
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'stroke-dasharray' to a number: -3.14
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'stroke-dasharray' to a number: 3.14
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'stroke-dasharray' to a number: calc(2 + 3)
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
Harness: the test ran to completion.
